requirements.txt
Flask==2.2.2
Flask-RESTful==0.3.9
Flask-JWT-Extended==4.4.0
jsonschema==4.4.0